UPDATE move recovery route and fix redirect
