--- app/api/auth/route.ts
@@ -1,31 +1,0 @@
-
-import { type EmailOtpType } from '@supabase/supabase-js'
-import { cookies } from 'next/headers'
-import { NextRequest, NextResponse } from 'next/server'
-// The client you created from the Server-Side Auth instructions
-import createClient from '@/supabase/clients/server'
-
-export async function GET(request: NextRequest) {
-  const { searchParams } = new URL(request.url)
-  const token_hash = searchParams.get('token_hash')
-  const type = searchParams.get('type') as EmailOtpType | null
-  const next = searchParams.get('next') ?? '/'
-  const redirectTo = request.nextUrl.clone()
-  redirectTo.pathname = next
-
-  if (token_hash && type) {
-    const supabase = createClient()
-
-    const { error } = await supabase.auth.verifyOtp({
-      type,
-      token_hash,
-    })
-    if (!error) {
-		NextResponse.json({msg: redirectTo})
-      return NextResponse.redirect(redirectTo)
-    }
-	else
-		return NextResponse.json({msg: error, redirectTo})
-  } else
-  		return NextResponse.json({msg: 'No token_hash or type'})
-}
